requirements.txt
crawl4ai 
google-cloud-aiplatform>=1.38.1
playwright>=1.40.0
python-dotenv>=1.0.0

